+++ app/views/beacons/beacons/_form.html.erb
