+++ app/views/beacons/beacons/edit.html.erb
