+++ app/views/beacons/beacons/index.html.erb
