+++ app/views/beacons/beacons/new.html.erb
