+++ app/views/beacons/beacons/show.html.erb
